Modify mobile to bigint

+++ corpmartapi/corpmart/models.py
@@ -56,7 +56,7 @@ class User(AbstractUser):
     username = None
     email = models.EmailField(_('email address'), unique=True)
     country_code = models.IntegerField(default=91)
-    mobile = models.IntegerField(unique=True)
+    mobile = models.BigIntegerField(unique=True)
     organisation_name = models.CharField(max_length=200, blank=True)
 
     USERNAME_FIELD = 'email'
